Add files via upload
Added two functions to get result sets back in the database file. Changed a few more things in the fxml and added controllers for all current buttons. 

Needs the where clauses in the SQL in the HandleAcceptedLoadButtonAction and the invited version need to be fixed to match the database.

The admin portion still needs to be done, I wasn't too sure how to approach it.

I was thinking of adding a second table to the view page to add more details, but wasn't sure what ones, lmk in discord what you guys think